Pointing to scripts online is harmful
The installation portion to the readme instructs the user to use a shell script provided by npm to install the package manager. This is harmful because it leaves open way too many ambiguities. A safer way would be to instruct the user to use a package manager to install npm (even though most Node installers provide npm by default now).