add option for client IP logging
Continuing work by @paked in #187, this adds the useful feature of
logging the remote client IP address. I personally use this feature in
Apache often at work, and it would be nice to have it here as well!

The addition is the `--log-ip` option, which adds the client IP to the
log message between the date and the request method and url.

Signed-off-by: Jade Thornton <jade@jmthornton.net>