Include changes form PR #923 with additional changes (#927)
* feat: add encoding even if .br/.gz specified (#923)

* Add handling of brotli/gzip pre-compressed files

Serving pre-compressed brotli/gzip files was not working before, because a file.br.br against the pre-compressed file.br check always returned false.

* Bump version from 14.1.1 to 14.1.2

* fix: always check the dot

This fixes an issue recently introduced where file extensions ending
with `br` are handled even when the dot does not immediately
preceded. This introduces erroneous an error when trying to access
files with other extensions ending with `br`, such as `.cbr`.

I meant to merge the previous PR with this fix included but forgot to
commit the suggestion I added in the PR review. This was merged to a
feature branch so the main branch was not affected.

* fix: put new behavior for .br/.gz behind a flag

Adds --force-content-encoding flag which must be specified to enable
the behavior of including the Content-Encoding header in the response
when a .br or .gz file is directly requested in the URL.

* test: --force-content-encoding flag

Tests whether or not Content-Encoding is present under different cases
with and without the --force-content-encoding flag set.

* doc: update manpage

* sync: package-lock.json

---------

Co-authored-by: Andreas Zeitler <andreas.zeitler@protonmail.com>