test: use ephemeral ports to avoid race conditions
Tests prior to this change use a mixture of hard-coded ports and ports
found using `portfinder`. Hard-coded ports cause tests to fail depending
on the conditions of the host environment they are running on. Ports
obtained using portfinder cause stochastic port collisions because tap
runs tests in parallel; two tests might find the same open port and both
try to listen on that port simultaneously.

This change updates several tests to use ephemeral. This is accomplished
by specifying the port as `0` in the call to `.listen()`:

> If port is omitted or is 0, the operating system will assign an
> arbitrary unused port, which can be retrieved by using
> server.address().port after the 'listening' event has been
> emitted.
source: https://nodejs.org/api/net.html#net_server_listen

Since the http server is `union` with a wrapper class,
`lib/http-server.js` is updated to inlcude the new method `address()`
which allows access to the ephemeral port number that was selected.