Do not allow CRLF in headers
While the underlying HTTP server does in fact throw an error when the
response will have invalid HTTP headers, this happens at runtime at the
time of the first request rather than immediately after the process is
started. This should be considered a fatal configuration error, causing
the process to exit immediately.