feat: add flag `--content-type`
Adds a flag to change the default Content-Type header value to something
other than application/octet-stream. This feature may not have the
demand that it would have before the internalization of ecstatic,
however contentType was already managed as an option and there just
wasn't a flag in front of it, so I figured I'd finish what was started.

Co-authored-by: howardroark <howardroark@users.noreply.github.com>