fix: fail fast on `--password` without `--user` (#937)
* fix: fail fast on `--password` without `--user`

When passing a password without a username, http-server will launch
successfully but the first request with an Authorization header will
crash the server. This behavior is very inconvenient. This commit makes
it fail fast instead so you know there's a problem before you log out of
SSH or walk out of the house or office. Also alised `--user` to
`--username` for added convenience.

Resolves #935

* doc: update documentation for `--username` flag